Tests/LibWeb/Text/expected/DOM/derived-interface-prototype.txt

derived class method present and invoked!
